+++ src/main/java/dev/krk/emulator/chipsekiz/opcodes/Op4XNN.java
@@ -5,7 +5,7 @@ package dev.krk.emulator.chipsekiz.opcodes;
  * Skips the next instruction if VX doesn't equal NN.
  */
 public class Op4XNN extends OpPXNN {
-    public Op4XNN(int vx, int imm) {
+    public Op4XNN(int vx, byte imm) {
         super(4, vx, imm);
     }
 }
